Set up Flask Backend
I've created a simple Flask application (main.py) to serve as the backend. The backend listens for POST requests on the /submit route and processes the data. I've also included Flask-CORS to allow communication between the frontend (HTML) and the backend (Python). I now generate poker hands on the backend and formatted them to be sent back as a response.