Two pair probability added
I've done probability calculators for all hand types. I've tested a lot of them as I've been going along but the next thing to do will be to get the web app working and able to handle card selections and display the appropriate probabilities.

Phew that was a lot of maths to get here!